+++ ricaun.Revit.Mvvm.Example/Properties/Resource.Designer.cs
@@ -19,7 +19,7 @@ namespace ricaun.Revit.Mvvm.Example.Properties {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "16.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class Resource {
@@ -61,11 +61,11 @@ namespace ricaun.Revit.Mvvm.Example.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to https://forested-image-generator.glitch.me/image/32/007700/circle.png.
+        ///   Looks up a localized string similar to Example.
         /// </summary>
-        internal static string Icon {
+        internal static string Text {
             get {
-                return ResourceManager.GetString("Icon", resourceCulture);
+                return ResourceManager.GetString("Text", resourceCulture);
             }
         }
     }