+++ ricaun.Revit.Mvvm/Input/AsyncRelayCommand.cs
@@ -110,11 +110,10 @@ namespace ricaun.Revit.Mvvm
                     finally
                     {
                         IsExecuting = false;
+                        RaiseCanExecuteChanged();
                     }
                 }
             }
-
-            RaiseCanExecuteChanged();
         }
     }
 }