+++ ricaun.Revit.Mvvm/Input/AsyncRelayCommand{T}.cs
@@ -141,11 +141,10 @@ namespace ricaun.Revit.Mvvm
                     finally
                     {
                         IsExecuting = false;
+                        RaiseCanExecuteChanged();
                     }
                 }
             }
-
-            RaiseCanExecuteChanged();
         }
     }
 }